BackgroundRemoteProcessorServer: rename header and source files to match class name
